scripts/requirements.txt
anthropic==0.40.0
supabase==2.10.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
resend==2.4.0
python-dateutil==2.9.0
tenacity==9.0.0
